+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.9.4](https://github.com/nx-dotnet/nx-dotnet/compare/v1.9.3...v1.9.4) (2022-03-01)
+
+### Bug Fixes
+
+- **core:** process project graph shouldn't throw if not a .NET project ([#390](https://github.com/nx-dotnet/nx-dotnet/issues/390)) ([4d51ea0](https://github.com/nx-dotnet/nx-dotnet/commit/4d51ea06d70307b8907a10adb26ca12ea7b0c886))
+
+## [1.9.3](https://github.com/nx-dotnet/nx-dotnet/compare/v1.9.2...v1.9.3) (2022-03-01)
+
+### Bug Fixes
+
+- **core:** infer-projects should work with nx daemon ([#383](https://github.com/nx-dotnet/nx-dotnet/issues/383)) ([acf47cf](https://github.com/nx-dotnet/nx-dotnet/commit/acf47cfaab6c6a125be6fb8205733c639977a203))
+
+## [1.9.2](https://github.com/nx-dotnet/nx-dotnet/compare/v1.9.1...v1.9.2) (2022-02-28)
+
+### Bug Fixes
+
+- **dotnet:** project with spaces do not bug out the command ([#379](https://github.com/nx-dotnet/nx-dotnet/issues/379)) ([dd4e16b](https://github.com/nx-dotnet/nx-dotnet/commit/dd4e16b109b51aba296c5faf9ce9d0c6fa4929c3))
+
 ## [1.9.1](https://github.com/nx-dotnet/nx-dotnet/compare/v1.9.0...v1.9.1) (2022-02-15)
 
 ### Bug Fixes